Restructure for api/core separation
