+++ its_hub/__init__.py
@@ -6,18 +6,16 @@ from importlib.metadata import version
 
 __version__ = version("its_hub")
 
-# Core abstractions - always available
-from .algorithms.bon import BestOfN
-
-# Core algorithms - always available
-from .algorithms.self_consistency import SelfConsistency
-from .base import (
+# Core - Algorithm implementations (always available)
+from its_hub.api import (
     AbstractLanguageModel,
     AbstractOutcomeRewardModel,
     AbstractProcessRewardModel,
     AbstractScalingAlgorithm,
     AbstractScalingResult,
 )
+from its_hub.core.algorithms.bon import BestOfN
+from its_hub.core.algorithms.self_consistency import SelfConsistency
 
 # Start with core exports
 __all__ = [  # noqa: RUF022
@@ -36,8 +34,9 @@ __all__ = [  # noqa: RUF022
 
 # Optional LM implementations - only available if [lm] extra is installed
 try:
-    from .lms import OpenAICompatibleLanguageModel, StepGeneration
-    from .reward_models import LLMJudge
+    from its_hub.core.lms.openai_lm import OpenAICompatibleLanguageModel
+    from its_hub.core.lms.step_generation import StepGeneration
+    from its_hub.core.reward_models.llm_judge import LLMJudge
 
     __all__.extend(["LLMJudge", "OpenAICompatibleLanguageModel", "StepGeneration"])
 except ImportError:
