+++ its_hub/api/reward_models/__init__.py
@@ -1,0 +1,9 @@
+"""Reward model abstract base classes."""
+
+from .orm import AbstractOutcomeRewardModel
+from .prm import AbstractProcessRewardModel
+
+__all__ = [
+    "AbstractOutcomeRewardModel",
+    "AbstractProcessRewardModel",
+]
