--- its_hub/base.py
@@ -1,180 +1,0 @@
-from abc import ABC, abstractmethod
-
-from .types import ChatMessage, ChatMessages
-
-
-class AbstractLanguageModel(ABC):
-    """
-    Abstract base class for language models.
-
-    Gateway integrators should implement this interface to use its_hub algorithms
-    with their existing LM infrastructure. Only async implementation is required.
-    """
-
-    @abstractmethod
-    async def agenerate(
-        self,
-        messages: list[ChatMessage] | list[list[ChatMessage]],
-        stop: str | None = None,
-        **kwargs,
-    ) -> dict | list[dict]:
-        """
-        Generate response(s) asynchronously.
-
-        Args:
-            messages: Single conversation or batch of conversations
-            stop: Optional stop sequence for generation
-            **kwargs: Additional model-specific parameters (tools, tool_choice, etc.)
-
-        Returns:
-            Single response dict or list of response dicts (for batched input)
-            Response dict format: {"role": "assistant", "content": "...", "tool_calls": [...]}
-        """
-        pass
-
-
-class AbstractScalingResult(ABC):
-    """
-    Abstract base class for algorithm results.
-
-    Algorithms return instances of this class when return_response_only=False.
-    """
-
-    @property
-    @abstractmethod
-    def the_one(self) -> dict:
-        """
-        Return the selected best response.
-
-        Returns:
-            The response message dict selected by the algorithm
-            Response dict format: {"role": "assistant", "content": "...", "tool_calls": [...]}
-        """
-        pass
-
-
-class AbstractScalingAlgorithm(ABC):
-    """
-    Abstract base class for inference-time scaling algorithms.
-
-    All algorithms (Self-Consistency, Best-of-N, etc.) implement this interface.
-    """
-
-    @abstractmethod
-    async def ainfer(
-        self,
-        lm: AbstractLanguageModel,
-        prompt_or_messages: str | list[ChatMessage] | ChatMessages,
-        budget: int,
-        return_response_only: bool = True,
-        tools: list[dict] | None = None,
-        tool_choice: str | dict | None = None,
-    ) -> dict | AbstractScalingResult:
-        """
-        Run inference asynchronously with the given language model and prompt.
-
-        Args:
-            lm: Language model instance implementing AbstractLanguageModel
-            prompt_or_messages: User prompt (string or structured messages)
-            budget: Computational budget (interpretation varies by algorithm)
-            return_response_only: If True, return just the selected response;
-                                   if False, return full result object
-            tools: Optional OpenAI-style tool definitions
-            tool_choice: Optional tool choice strategy ("auto", "none", or specific tool)
-
-        Returns:
-            Selected response dict (if return_response_only=True) or
-            AbstractScalingResult instance with full details
-            Response dict format: {"role": "assistant", "content": "...", "tool_calls": [...]}
-        """
-        pass
-
-    def infer(
-        self,
-        lm: AbstractLanguageModel,
-        prompt_or_messages: str | list[ChatMessage] | ChatMessages,
-        budget: int,
-        return_response_only: bool = True,
-        tools: list[dict] | None = None,
-        tool_choice: str | dict | None = None,
-    ) -> dict | AbstractScalingResult:
-        """
-        Run inference synchronously with the given language model and prompt.
-
-        Default implementation wraps ainfer() using asyncio.run().
-        """
-        import asyncio
-
-        return asyncio.run(
-            self.ainfer(
-                lm, prompt_or_messages, budget, return_response_only, tools, tool_choice
-            )
-        )
-
-
-class AbstractOutcomeRewardModel(ABC):
-    """
-    Abstract base class for outcome reward models and judge models.
-
-    This class supports both traditional reward models and LLM-based judge models
-    that evaluate conversation outcomes and quality.
-    """
-
-    @abstractmethod
-    def score(
-        self,
-        messages: list[list[dict]] | list[dict],
-        **kwargs,
-    ) -> list[float] | float:
-        """
-        Score conversations synchronously.
-
-        Args:
-            messages: Single conversation or batch of conversations
-                Single: list[dict] (one conversation)
-                Batch: list[list[dict]] (multiple conversations)
-            **kwargs: Additional model-specific parameters
-
-        Returns:
-            Single score (float) or list of scores (list[float])
-            Higher score = better response
-        """
-        pass
-
-    async def ascore(
-        self,
-        messages: list[list[dict]] | list[dict],
-        **kwargs,
-    ) -> list[float] | float:
-        """
-        Score conversations asynchronously.
-
-        Default implementation raises NotImplementedError.
-        Override this for async-compatible reward models (e.g., LLM judges).
-        """
-        raise NotImplementedError(
-            f"{self.__class__.__name__} does not implement async scoring. "
-            "Override ascore() to support async scoring."
-        )
-
-
-class AbstractProcessRewardModel(ABC):
-    """abstract base class for process reward models"""
-
-    @abstractmethod
-    async def ascore(
-        self,
-        prompt_or_messages: str | list[ChatMessage] | ChatMessages,
-        steps: list[str],
-    ) -> list[float]:
-        """score steps asynchronously"""
-        pass
-
-    @abstractmethod
-    def score(
-        self,
-        prompt_or_messages: str | list[ChatMessage] | ChatMessages,
-        steps: list[str],
-    ) -> list[float]:
-        """score steps synchronously"""
-        pass
