+++ its_hub/core/algorithms/particle_gibbs.py
@@ -5,14 +5,15 @@ from enum import Enum
 
 import numpy as np
 
-from its_hub.base import (
+from its_hub.api import (
     AbstractLanguageModel,
     AbstractProcessRewardModel,
     AbstractScalingAlgorithm,
     AbstractScalingResult,
+    ChatMessage,
+    ChatMessages,
 )
-from its_hub.lms import StepGeneration
-from its_hub.types import ChatMessage, ChatMessages
+from its_hub.core.lms.step_generation import StepGeneration
 
 
 @dataclass
